Made minor changes in the code
As I have already said, I'm just getting into Python and I don't know much and so I haven't changed the functionality in any way, just the internal workings of the program.

So, this modification makes use of elif and else to improve the detection of voice recognition errors - specifically, when it hears "sub tractor" instead of "subtraction" ie. mishearing.

Along with that, I replaced an "ok" in the code with "hello", it won't change anything, I promise, it's just for fun.

Not to mention that I had changed the order of some of the instructions that appear at the start.

But the most important part is that PYTTSX3 IS NOT REQUIRED ANYMORE (at least for now)!


Thanks and I hope you will find this little piece of code helpful.